Extract BlueMessageBar
