+++ portal/src/graphql/portal/FeatureDisabledMessageBar.tsx
@@ -1,13 +1,8 @@
 import React, { useMemo } from "react";
-import {
-  MessageBar,
-  IMessageBarProps,
-  useTheme,
-  PartialTheme,
-  ThemeProvider,
-} from "@fluentui/react";
-import { FormattedMessage, Values } from "@oursky/react-messageformat";
+import { IMessageBarProps } from "@fluentui/react";
 import { useParams } from "react-router-dom";
+import { FormattedMessage, Values } from "@oursky/react-messageformat";
+import BlueMessageBar from "../../BlueMessageBar";
 
 export interface FeatureDisabledMessageBarProps extends IMessageBarProps {
   messageID: string;
@@ -19,16 +14,6 @@ const FeatureDisabledMessageBar: React.VFC<FeatureDisabledMessageBarProps> =
     const { messageID, messageValues, ...rest } = props;
     const { appID } = useParams() as { appID: string };
 
-    const theme = useTheme();
-    const newTheme: PartialTheme = {
-      semanticColors: {
-        messageText: theme.palette.themePrimary,
-        messageLink: theme.semanticColors.link,
-        messageLinkHovered: theme.semanticColors.linkHovered,
-        infoIcon: theme.palette.themePrimary,
-      },
-    };
-
     const values = useMemo(() => {
       return {
         planPagePath: `/project/${appID}/billing`,
@@ -37,29 +22,9 @@ const FeatureDisabledMessageBar: React.VFC<FeatureDisabledMessageBarProps> =
     }, [appID, messageValues]);
 
     return (
-      <ThemeProvider theme={newTheme}>
-        <MessageBar
-          {...rest}
-          styles={{
-            root: {
-              background: theme.palette.themeLighter,
-            },
-            innerText: {
-              fontSize: "14px",
-              lineHeight: "20px",
-              a: {
-                fontWeight: 600,
-                whiteSpace: "nowrap",
-              },
-            },
-            icon: {
-              lineHeight: "20px",
-            },
-          }}
-        >
-          <FormattedMessage id={messageID} values={values} />
-        </MessageBar>
-      </ThemeProvider>
+      <BlueMessageBar {...rest}>
+        <FormattedMessage id={messageID} values={values} />
+      </BlueMessageBar>
     );
   };
